requirements.txt
# Langchain core 
langchain
langchain-core

#OpenAI Integration
langchain-openai
openai

#Anthropic Integration
langchain-anthropic

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#HuggingFace Integration
langchain-huggingface
transformers
huggingface-hub

#Environment Variables
python-dotenv

#Machine Learning
scikit-learn
numpy